+++ include/config.h
@@ -78,6 +78,7 @@ public:
         this->monitorFiles = true;
         this->tilesetCheckerboardFill = true;
         this->theme = "default";
+        this->wildMonChartTheme = "";
         this->textEditorOpenFolder = "";
         this->textEditorGotoLine = "";
         this->paletteEditorBitDepth = 24;
@@ -128,6 +129,7 @@ public:
     bool monitorFiles;
     bool tilesetCheckerboardFill;
     QString theme;
+    QString wildMonChartTheme;
     QString textEditorOpenFolder;
     QString textEditorGotoLine;
     int paletteEditorBitDepth;
@@ -137,6 +139,7 @@ public:
     QDateTime lastUpdateCheckTime;
     QVersionNumber lastUpdateCheckVersion;
     QMap<QUrl, QDateTime> rateLimitTimes;
+    QByteArray wildMonChartGeometry;
 
 protected:
     virtual QString getConfigFilepath() override;
@@ -192,11 +195,13 @@ enum ProjectIdentifier {
     define_obj_event_count,
     define_min_level,
     define_max_level,
+    define_max_encounter_rate,
     define_tiles_primary,
     define_tiles_total,
     define_metatiles_primary,
     define_pals_primary,
     define_pals_total,
+    define_tiles_per_metatile,
     define_map_size,
     define_mask_metatile,
     define_mask_collision,
@@ -216,6 +221,7 @@ enum ProjectIdentifier {
     define_map_section_prefix,
     define_map_section_empty,
     define_map_section_count,
+    define_species_prefix,
     regex_behaviors,
     regex_obj_event_gfx,
     regex_items,
@@ -230,7 +236,6 @@ enum ProjectIdentifier {
     regex_sign_facing_directions,
     regex_trainer_types,
     regex_music,
-    regex_species,
 };
 
 enum ProjectFilePath {
@@ -279,6 +284,7 @@ enum ProjectFilePath {
     global_fieldmap,
     fieldmap,
     initial_facing_table,
+    wild_encounter,
     pokemon_icon_table,
     pokemon_gfx,
 };